requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
sqlalchemy
xgboost
prophet
streamlit
plotly
rich
python-dotenv
praw
requests
transformers
torch
scikit-learn
python-dateutil
aiohttp
asyncio-throttle
ta
joblib
pytest
